avoid webNavigation permission
Produces a warning because this gets URLs for any host, even without this being an active tab.
Use `chrome.tabs.onUpdated` with activeTab instead. This allows me to continue checking the URL, until the user navigates to another host.

Note: if user enables extension on host A, then navigates A -> B -> A, it will show wrong state on 2nd time on A